requirements.txt
python-telegram-bot==21.4
feedparser==6.0.10
apscheduler==3.10.4
